Rename to state.py
